feat: split calendar view
