+++ src/components/calendar-view/calendar-view.en.md
@@ -1,42 +1,26 @@
-# Calendar <Experimental></Experimental>
+# CalendarView <Experimental></Experimental>
 
 Used to select a date or date range.
 
-## When to Use
-
-When the user needs to enter a date, he can select it in the pop-up date panel.
+CalendarView is the content area of [Calendar](/components/calendar).
 
 ## Demos
 
+Only the simplest content area is shown here, and other more usages can be consulted [Calendar](/zh/components/calendar)
+
 <code src="./demos/demo1.tsx"></code>
 
-<code src="./demos/demo2.tsx"></code>
-
-<code src="./demos/demo3.tsx"></code>
-
-<code src="./demos/demo4.tsx"></code>
-
-## Calendar
+## CalendarView
 
 ### Props
 
 | Name | Description | Type | Default | Version |
 | --- | --- | --- | --- | --- |
-| visible | To show or hide the Cclendar | `boolean` | `true` |
-| confirmText | The text of confirm button | `string` | `Confirm` |
-| popupClassName | The custom class name of the popup | `string` | - |
-| popupStyle | The custom style of the popup | `React.CSSProperties` | - |
-| popupBodyStyle | The custom style of the popup body | `React.CSSProperties` | - |
-| forceRender | Render content forcely，When ref is passed，always be true | `boolean` | `false` |
-| closeOnMaskClick | Whether to close after clicking the mask layer | `boolean` | `true` |
-| onClose | Triggered when closed | `() => void` | - |
-| onMaskClick | Triggered when the mask is clicked | `() => void` | - |
 | allowClear | Whether to allow clearing after another click. | `boolean` | `true` |
 | defaultValue | The default selected date or date range. | Same as `value` prop. | - |
 | max | Maximum value of a selectable range. | `Date` | - |
 | min | Minimum value of a selectable range. | `Date` | - | - |
 | onChange | Trigger when selected date changes. | `(val: Date \| null) => void` when selection mode is "single". `(val: [Date, Date] \| null) => void` when selection mode is "range". | - |
-| onConfirm | Trigger when confirm button is clicked. | `(val: Date \| null) => void` when selection mode is "single"，`(val: [Date, Date] \| null) => void` when selection mode is "range" | - |
 | renderTop | The top information of date render function. | `(date: Date) => ReactNode \| null \| undefined` | - |
 | renderBottom | The bottom information of date render function. | `(date: Date) => ReactNode \| null \| undefined` | - |
 | selectionMode | The selection mode. Disable selection when this prop is not set. | `'single' \| 'range'` | - |
