+++ src/components/calendar-view/demos/demo1.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import { Calendar } from 'antd-mobile'
+import { CalendarView } from 'antd-mobile'
 import { DemoBlock } from 'demos'
 
 export default () => {
   return (
-    <DemoBlock title='不使用 Popup'>
-      <Calendar usePopup={false} />
+    <DemoBlock title='只展示内容区域'>
+      <CalendarView />
     </DemoBlock>
   )
 }
