+++ src/components/calendar-view/index.ts
@@ -1,0 +1,6 @@
+import './calendar-view.less'
+import { CalendarView } from './calendar-view'
+
+export type { CalendarViewProps, CalendarViewRef } from './calendar-view'
+
+export default CalendarView
