+++ src/components/calendar/demos/demo2.tsx
@@ -22,15 +22,9 @@ export default () => {
         自定义标题
         <Calendar
           title='请选择日期'
-          popupProps={{
-            visible: visible1,
-            onMaskClick: () => {
-              setVisible1(false)
-            },
-            onClose: () => {
-              setVisible1(false)
-            },
-          }}
+          visible={visible1}
+          onClose={() => setVisible1(false)}
+          onMaskClick={() => setVisible1(false)}
         />
       </List.Item>
       <List.Item
@@ -41,15 +35,9 @@ export default () => {
         自定义确认文案
         <Calendar
           confirmText='确认选择'
-          popupProps={{
-            visible: visible2,
-            onMaskClick: () => {
-              setVisible2(false)
-            },
-            onClose: () => {
-              setVisible2(false)
-            },
-          }}
+          visible={visible2}
+          onClose={() => setVisible2(false)}
+          onMaskClick={() => setVisible2(false)}
         />
       </List.Item>
       <List.Item
@@ -60,15 +48,9 @@ export default () => {
         周一作为每周的第一天
         <Calendar
           weekStartsOn='Monday'
-          popupProps={{
-            visible: visible3,
-            onMaskClick: () => {
-              setVisible3(false)
-            },
-            onClose: () => {
-              setVisible3(false)
-            },
-          }}
+          visible={visible3}
+          onClose={() => setVisible3(false)}
+          onMaskClick={() => setVisible3(false)}
         />
       </List.Item>
       <List.Item
@@ -81,15 +63,9 @@ export default () => {
           min={min}
           max={max}
           selectionMode='range'
-          popupProps={{
-            visible: visible4,
-            onMaskClick: () => {
-              setVisible4(false)
-            },
-            onClose: () => {
-              setVisible4(false)
-            },
-          }}
+          visible={visible4}
+          onClose={() => setVisible4(false)}
+          onMaskClick={() => setVisible4(false)}
         />
       </List.Item>
     </List>
