+++ src/components/calendar/index.ts
@@ -1,4 +1,3 @@
-import './calendar.less'
 import { Calendar } from './calendar'
 
 export type { CalendarProps, CalendarRef } from './calendar'
