+++ src/index.ts
@@ -14,6 +14,11 @@ export { default as Button } from './components/button'
 export type { ButtonProps, ButtonRef } from './components/button'
 export { default as Calendar } from './components/calendar'
 export type { CalendarProps, CalendarRef } from './components/calendar'
+export { default as CalendarView } from './components/calendar-view'
+export type {
+  CalendarViewProps,
+  CalendarViewRef,
+} from './components/calendar-view'
 export { default as CapsuleTabs } from './components/capsule-tabs'
 export type {
   CapsuleTabsProps,
